FE/Qt: Build warning fix for r98178.
svn:sync-xref-src-repo-rev: r98180

+++ src/VBox/Frontends/VirtualBox/src/runtime/UIIndicatorsPool.cpp
@@ -1,4 +1,4 @@
-/* $Id: UIIndicatorsPool.cpp 54147 2015-02-11 16:16:36Z sergey.dubov@oracle.com $ */
+/* $Id: UIIndicatorsPool.cpp 54148 2015-02-11 16:27:33Z sergey.dubov@oracle.com $ */
 /** @file
  * VBox Qt GUI - UIIndicatorsPool class implementation.
  */
@@ -395,7 +395,7 @@ private:
         QStringList ipList, macList;
         if (fPropsValid)
         {
-            const int cAdapters = RT_MIN(strCount.toInt(), m_cMaxNetworkAdapters);
+            const int cAdapters = RT_MIN(strCount.toInt(), (int)m_cMaxNetworkAdapters);
             for (int i = 0; i < cAdapters; ++i)
             {
                 ipList << machine.GetGuestPropertyValue(QString("/VirtualBox/GuestInfo/Net/%1/V4/IP").arg(i));
@@ -457,7 +457,7 @@ private:
     /** Holds the auto-update timer instance. */
     QTimer *m_pTimerAutoUpdate;
     /** Holds the maximum amount of the network adapters. */
-    int m_cMaxNetworkAdapters;
+    ulong m_cMaxNetworkAdapters;
 };
 
 /** UISessionStateStatusBarIndicator extension for Runtime UI: USB indicator. */
